Gracefully handle expired access token
@andreoda discovered that some playlists weren't updating:
https://github.com/mackorone/spotify-playlist-archive/pull/302

Turns out, the scraper has been silently failing to fetch playlists for
more than three months (ouch). The problem is that the Spotify access
token expires after an hour and all subsequent fetches will fail. At
some point, scraping started to take longer than an hour, and rather
than raising an exception, we simply swallow the error. The swallowing
logic originally intended to handle missing playlists; "expired access
token" errors simply got caught in the crossfire.

This commit fixes the issue by copying some code from a related project,
mackorone/spotify-playlist-publisher. This new code has more nuance when
it comes to error handling. In particular, when an access token expires,
we just fetch a new one. And when we fail to fetch a playlist because it
doesn't exist, we log it and move on.